Moved GameInput patches to ClientBase, these need to be initialized before game start.
